# Optimize compilation speed by 42%